+++ definitions.h
@@ -1,0 +1,62 @@
+#include <string>
+#include <vector>
+
+using namespace std;
+
+enum MENU_CHOICES { EXIT = 'X', PROG_DESC = 'P', EVALUATE = 'E' };
+typedef int token_type;
+typedef int type_data;
+typedef char choice;
+
+/**
+ * `token` type that is what the lexer returns.
+ *
+ * these tokens correspond to terminal types in the grammar.
+ */
+enum TOKEN_TYPES { NUMBER, PRECEDENCE_1, PRECEDENCE_2 };
+struct token {
+  token_type type;
+  string value;
+};
+
+/**
+ * `node` type that can be either `Numeric `or `Operator`
+ *
+ * No methods because only limited to procedural code.
+ *
+ * `Numeric` types hold a value but no children nodes.
+ * `Operator` types hold a value, and children nodes.
+ */
+enum NODE_TYPES { NUMERIC, OPERATOR };
+struct node {
+  type_data type;
+  int value;
+  node &left;
+  node &right;
+};
+
+/**
+ * Menu interface, returns a choice value based on what user inputs.
+ *
+ * @return menu choice
+ */
+choice menu(void);
+
+/**
+ * Prints out the program description to `stdout`.
+ */
+void programDescription(void);
+
+/**
+ * Enter the loop for evaluating expressions.
+ * Exit back to `main` on 'X' or invalid expression.
+ */
+void evaluateLoop(void);
+
+/**
+ * Converts the string expression to its corresponding `token` structs.
+ *
+ * @param expression The expression to be lexed (filtered for spaces)
+ * @ret the tokens for parsing
+ */
+vector<token> lexer(const string expression);
